fixed trailer in my library, added constants file, all constants should be there
