+++ src/js/constants.js
@@ -1,0 +1,3 @@
+export const API = 'https://api.themoviedb.org/3/';
+export const KEY = '9068359f92c010fa6a3cf763f10a0606';
+export const MEDIA_TYPE = 'movie';
