+++ src/js/trailer.js
@@ -1,4 +1,28 @@
-import { getMovieTrailer } from './popular';
+import axios from 'axios';
+import { KEY, API, MEDIA_TYPE } from './constants';
+
+export const getMovieTrailer = async movieId => {
+  try {
+    const searchParams = new URLSearchParams({
+      api_key: KEY,
+    });
+
+    const response = await axios.get(
+      `${API}${MEDIA_TYPE}/${movieId}/videos?${searchParams}`
+    );
+    console.log(
+      'here',
+      `${API}${MEDIA_TYPE}/${movieId}/videos?${searchParams}`
+    );
+    if (response.status !== 200) {
+      throw new Error(response.status);
+    }
+
+    return response.data;
+  } catch (error) {
+    // console.log('Підставити картинку, сервер терміново недоступний');
+  }
+};
 
 export const trailerButtonRef = document.getElementById('trailer');
 const trailerErrorContainer = document.getElementById('trailer-error');
