+++ src/library.html
@@ -19,6 +19,7 @@
     <!-- <include src="./partials/card-film-list.html"></include>     -->
     <include src="./partials/pagination.html"></include>
     <include src="./partials/modal-team.html"></include>
+    <include src="./partials/spinner.html"></include>
     <include src="./partials/footer.html"></include>
   </body>
 </html>
