+++ src/library.js
@@ -3,5 +3,5 @@ import './js/btn-scroll';
 import './js/watched';
 import './js/queue';
 import './js/local-storage-set';
-
+import './js/trailer';
 import './js/open-modal-library';
